Fix Abe always defaulting to NO_CLOB
Using the MySQL driver, I noticed Abe would no longer select LONGTEXT as
CLOB replacement. It turns out there was a mix of format string and SQL
parameter, but both values were passed to the SQL driver which obviously
didn't like it: This would silently default to BUG_NO_CLOB unless
logging was set to INFO or higher:

2015-01-28 00:47:50,599 [1925:MainThread] Abe.SqlAbstraction.sql INFO -
  EXCEPTION: not all arguments converted during string formatting

Catching "Exception" is one way used to tell a data type didn't work, so
this error was ignored as part of the logic.