Mempool transactions loading fixes
Mempool loading used to be a one-time process prone to errors. This
commit fixes several things:

1. Don't load mempool by default. Abe is not yet able to clean
transactions that never make it to blocks, leading to a signigicant
number of stale, unreferenced transactions in the database.

2. When mempool loading is enabled, loop forever until a new block is
found.

3. Gracefully ignore transactions cleaned up from the mempool.

4. Speed up mempool tx check. We used to fetch the full tx merely to
know if it existed, now we only load the tx data when we need it.