Add default chain parameter
Chain.create needs to be called with a chain name, so adding a default
one.

This doesn't seem to do much actually - The only thing is fixes for me
is looking up mempool or stale transactions which aren't linked to any
block (so not any chain neither). Therefore I'm not adding it as an
abe.conf setting.