Fix rounding errors in block's ss_destroyed and ss stats
block_ss_destroyed and block_satoshi_seconds relied on SQL
multiplication. It was noticed on MySQL that because the nTime value
part of the expression was passed as decimal.Decimal(n) to the MySQL
driver it ended up using floating point math instead of precision math.

The fix could have been to cast nTime to integer, however to ensure the
most consistent and correct behavior regardless of the driver the
multiplication is now done in Python.